skip getStarted from step 4
